# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mefali API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/vendeur/prestataires": {
      "get": {
        "tags": [
          "vendeur"
        ],
        "summary": "Prestataires que ce compte pilote (rattachements du cycle VND).",
        "operationId": "mes_prestataires",
        "responses": {
          "200": {
            "description": "Prestataires rattachés, plus ancien rattachement d'abord — l'app pilote le premier au MVP (aucune sélection de site n'existe, FR-019).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrestatairePilotable"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Rôle vendeur requis.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/vendeur/prestataires/{id}/articles": {
      "get": {
        "tags": [
          "vendeur"
        ],
        "summary": "Catalogue COMPLET du prestataire piloté (ruptures, retirés, verrou admin).",
        "operationId": "mes_articles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Prestataire piloté.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Catalogue de pilotage — les retirés en tête de leur groupe, remise possible sans ressaisie (FR-055).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleVendeur"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Rôle vendeur absent, non rattaché, ou prestataire non agréé (trois codes distincts — R11).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "vendeur"
        ],
        "summary": "Ajoute un article au catalogue (V2 — « + Ajouter un article »).",
        "operationId": "creer_article",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Prestataire piloté.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreerArticleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Article créé, disponible par défaut (FR-020). Émet `article.cree`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVendeur"
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Refus de pilotage (trois codes distincts).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "422": {
            "description": "Prix barré ≤ prix (FR-023), montant négatif, nom vide.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/vendeur/prestataires/{id}/articles/{article_id}": {
      "put": {
        "tags": [
          "vendeur"
        ],
        "summary": "Modifie nom / prix / prix barré / étiquette (fiche article V2).",
        "operationId": "modifier_article",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Prestataire piloté.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "description": "Article.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifierArticleDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Modifié — `prix_barre_unites: null` retire la promotion EXPLICITEMENT. Un montant déjà FIGÉ ne bouge jamais (SC-005). Émet `article.modifie`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVendeur"
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Refus de pilotage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "404": {
            "description": "Article inconnu ou retiré du catalogue.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "422": {
            "description": "Prix barré qui deviendrait ≤ prix : l'opération ÉCHOUE, la promotion n'est pas retirée en silence (FR-023).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/vendeur/prestataires/{id}/articles/{article_id}/photo": {
      "post": {
        "tags": [
          "vendeur"
        ],
        "summary": "Dépose/remplace la photo de l'article (multipart, ≤ 5 Mo).",
        "operationId": "photo_article",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Prestataire piloté.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "description": "Article.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/DepotPhoto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Photo remplacée (clé neuve, l'ancienne purgée après commit).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVendeur"
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Refus de pilotage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "404": {
            "description": "Article inconnu ou retiré.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "422": {
            "description": "Type refusé ou fichier trop volumineux.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/vendeur/prestataires/{id}/articles/{article_id}/remise": {
      "post": {
        "tags": [
          "vendeur"
        ],
        "summary": "Remet un article retiré au catalogue, sans ressaisie (FR-055).",
        "operationId": "remettre_article",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Prestataire piloté.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "description": "Article.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remis : il revient avec son historique et sa disponibilité telle qu'elle était. Émet `article.remis_au_catalogue`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVendeur"
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Refus de pilotage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "404": {
            "description": "Article inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/vendeur/prestataires/{id}/articles/{article_id}/retrait": {
      "post": {
        "tags": [
          "vendeur"
        ],
        "summary": "Retire l'article du catalogue — RÉVERSIBLE (FR-055).",
        "operationId": "retirer_article",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Prestataire piloté.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "article_id",
            "in": "path",
            "description": "Article.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retiré : plus servi, plus commandable, plus signalable — la ligne subsiste. Émet `article.retire_du_catalogue`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleVendeur"
                }
              }
            }
          },
          "401": {
            "description": "Session absente, invalide ou révoquée.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "403": {
            "description": "Refus de pilotage.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          },
          "404": {
            "description": "Article inconnu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErreurApi"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ArticleVendeur": {
        "type": "object",
        "description": "Article du catalogue de PILOTAGE (écran V2) : ruptures, retirés et verrou\nadmin visibles — contrairement à la consultation publique.",
        "required": [
          "id",
          "nom",
          "prix_unites",
          "devise",
          "disponible",
          "rupture_admin",
          "retire"
        ],
        "properties": {
          "categorie_interne": {
            "type": [
              "string",
              "null"
            ],
            "description": "Étiquette libre de regroupement."
          },
          "devise": {
            "type": "string",
            "description": "Code ISO 4217 (posé par le serveur — R13)."
          },
          "disponible": {
            "type": "boolean",
            "description": "Faux = rupture."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identifiant."
          },
          "nom": {
            "type": "string",
            "description": "Nom."
          },
          "photo_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "URL présignée de la photo (TTL 10 min)."
          },
          "prix_barre_unites": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Prix barré (strictement supérieur — FR-023)."
          },
          "prix_unites": {
            "type": "integer",
            "format": "int64",
            "description": "Prix courant, entier en unités mineures."
          },
          "retire": {
            "type": "boolean",
            "description": "Retiré du catalogue — remise possible sans ressaisie (FR-055)."
          },
          "rupture_admin": {
            "type": "boolean",
            "description": "Rupture posée par l'Admin — la bascule vendeur sera refusée (FR-041)."
          },
          "source_derniere_bascule": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SourceBascule",
                "description": "Source de la dernière bascule (FR-037)."
              }
            ]
          }
        }
      },
...
      "CreerArticleDto": {
        "type": "object",
        "description": "Création d'un article (disponible par défaut — FR-020).",
        "required": [
          "nom",
          "prix_unites"
        ],
        "properties": {
          "categorie_interne": {
            "type": [
              "string",
              "null"
            ],
            "description": "Étiquette libre de regroupement."
          },
          "nom": {
            "type": "string",
            "description": "Nom.",
            "maxLength": 120,
            "minLength": 1
          },
          "prix_barre_unites": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Prix barré optionnel (strictement supérieur — FR-023)."
          },
          "prix_unites": {
            "type": "integer",
            "format": "int64",
            "description": "Prix courant, entier en unités mineures — la devise est POSÉE PAR LE\nSERVEUR depuis la zone (constitution III).",
            "minimum": 0
          }
        }
      },
...
      "DepotPhoto": {
        "type": "object",
        "description": "Photo envoyée en multipart (≤ 5 Mo, jpeg/png/webp).",
        "required": [
          "fichier"
        ],
        "properties": {
          "fichier": {
            "type": "string",
            "format": "binary",
            "description": "La photo."
          }
        }
      },
...
      "ErreurApi": {
        "type": "object",
        "description": "Corps d'erreur du contrat — `{ code, message_cle }`.",
        "required": [
          "code",
          "message_cle"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Code stable, exploitable par le client."
          },
          "message_cle": {
            "type": "string",
            "description": "Clé i18n fr — aucune chaîne UI en dur (constitution VII)."
          }
        }
      },
...
      "ModifierArticleDto": {
        "type": "object",
        "description": "Modification partielle — champ ABSENT = inchangé, `null` EXPLICITE = effacé\n(c'est ainsi qu'on retire une promotion : `prix_barre_unites: null`).",
        "properties": {
          "categorie_interne": {
            "type": [
              "string",
              "null"
            ],
            "description": "Nouvelle étiquette — `null` l'efface."
          },
          "nom": {
            "type": [
              "string",
              "null"
            ],
            "description": "Nouveau nom.",
            "maxLength": 120,
            "minLength": 1
          },
          "prix_barre_unites": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Nouveau prix barré — `null` retire la promotion EXPLICITEMENT (jamais\nen silence : un prix barré devenu ≤ prix fait échouer l'opération)."
          },
          "prix_unites": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Nouveau prix courant.",
            "minimum": 0
          }
        }
      },
...
      "PrestatairePilotable": {
        "type": "object",
        "description": "Prestataire pilotable par le compte (résumé — l'app prend le premier).",
        "required": [
          "id",
          "nom",
          "statut",
          "boutique"
        ],
        "properties": {
          "boutique": {
            "$ref": "#/components/schemas/EtatEffectifBoutique",
            "description": "État effectif de la boutique."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identifiant."
          },
          "nom": {
            "type": "string",
            "description": "Nom public."
          },
          "statut": {
            "$ref": "#/components/schemas/StatutPrestataire",
            "description": "Cycle de vie — `suspendu` : l'app affiche le refus, le rôle est intact."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}